tests: split the CLI integration harness by behavior
Replace the monolithic `gestalt/cli/tests/integration.rs` file with
behavior-focused integration suites and a shared harness module. Runtime
code stays unchanged; this only reorganizes how the existing CLI and
transport contracts are exercised.

Rust interfaces:
- no production Rust API changes in this commit
- integration suites now live in `cli/tests/auth_flow.rs`, `config_resolution.rs`, `invoke_contract.rs`, `plugins_connect.rs`, and `tokens_cli.rs`
- shared harness helpers now live in `cli/tests/support/mod.rs`

stdin/stdout interfaces pinned by the split:
- `gestalt auth status --format json`
- `printf '{"count":42}' | gestalt invoke test_svc do_thing --input-file - --select result.id`
- `gestalt plugins connect widget_metrics`
- `gestalt plugins list`

Testing:
- `cargo test -p gestalt`
